refactor: User 서비스 CQRS 패턴 적용
