+++ src/main/java/store/lastdance/service/user/UserV2Service.java
@@ -11,18 +11,12 @@ import java.util.UUID;
 
 public interface UserV2Service {
 
-    User findByActiveUser(UUID userid);
-    User findByUserId(UUID userId);
-    UserResponseDTO getUserWithProfileImage(UUID userId);
-
     User updateMyInfo(UUID userId, UserUpdateRequestDTO requestDTO);
+
     UserResponseDTO updateProfileImage(UUID userid, MultipartFile file);
+
     UserResponseDTO deleteProfileImage(UUID userid);
 
-    boolean isNicknameAvailable(UUID userId, String nickname);
-
-    // 삭제 = 계정 비활성화 (소프트 딜리트)
     void deactivateUser(UUID userId, HttpServletRequest request, HttpServletResponse response);
 
-    void validateUserExists(UUID userId);
 }
